Rename TestableVoiceChannel.swift to MockVoiceChannel
